feat: add tracing availability check, update docs
