+++ assembly/src/ast/nodes/mod.rs
@@ -285,10 +285,8 @@ pub enum Instruction {
     Breakpoint,
     Debug(DebugOptions),
 
-    // ----- emit instruction ---------------------------------------------------------------------
+    // ----- event decorators ---------------------------------------------------------------------
     Emit(u32),
-
-    // ----- trace instruction --------------------------------------------------------------------
     Trace(u32),
 }
 
@@ -548,10 +546,8 @@ impl fmt::Display for Instruction {
             Self::Breakpoint => write!(f, "breakpoint"),
             Self::Debug(options) => write!(f, "debug.{options}"),
 
-            // ----- emit instruction -------------------------------------------------------------
+            // ----- event decorators -------------------------------------------------------------
             Self::Emit(value) => write!(f, "emit.{value}"),
-
-            // ----- trace instruction ------------------------------------------------------------
             Self::Trace(value) => write!(f, "trace.{value}"),
         }
     }
