+++ assembly/src/ast/nodes/serde/mod.rs
@@ -257,10 +257,8 @@ pub enum OpCode {
     // ----- debugging ----------------------------------------------------------------------------
     Debug = 220,
 
-    // ----- emit ---------------------------------------------------------------------------------
+    // ----- event decorators ---------------------------------------------------------------------
     Emit = 221,
-
-    // ----- trace --------------------------------------------------------------------------------
     Trace = 222,
 
     // ----- control flow -------------------------------------------------------------------------
